+++ R/00.R
@@ -7,12 +7,20 @@
 #' Configuration
 #' \describe{
 #'  \item{root_url}{Root of the pubweb database site}
+#'  \item{cache_dir}{Directory where to place cache files (loaded from server)}
+#'  \item{curl_verbose}{Show verbose output when loading data from server}
+#'  \item{curl_insecure}{Disable certificate check when loaded data from server}
+#'  \item{records_page_size}{Number of records to load from the server at each step}
 #'  \item{teams}{vector of int, List of teams with numerical code as value, and team label as name}
 #' }
 config = rlang::new_environment(list(
   cache_dir = NULL,
   root_url = "https://pubweb.iplesp.fr",
-  teams = c('SUMO'=1L,'CLEPIVIR'=2L,'THERAVIR'=4L,'EPAR'=5L,'PEPITES'=6L,'ESSMA'=7L,'NEMESIS'=10L, 'CIPHOD'=11L)
+  curl_verbose=FALSE,
+  curl_insecure=FALSE,
+  records_page_size=2000,
+  default_ttl=3600L*3L,
+  teams = c()
 ))
 
 set_root_url = function(value) {
@@ -31,6 +39,7 @@ set_teams = function(value) {
   if(is.null(names(value))) {
     rlang::abort("Values for teams must be named")
   }
+  config$teams = value
 }
 
 set_cache_dir = function(value) {
@@ -51,12 +60,17 @@ pubweb_set_config = function(...) {
   if(is.null(names(oo))) {
     rlang::abort("Each parameter must be named")
   }
+
+  set_config = \(name, value) {
+    config[[name]] = value
+  }
+
   for(n in names(oo)) {
     switch(n,
       root_url=set_root_url(oo[[n]]),
       team=set_teams(oo[[n]]),
       cache_dir=set_cache_dir(oo[[n]]),
-      rlang::abort(paste("Unknown config", sQuote(n)))
+      set_config(n, oo[[n]])
     )
   }
 }