+++ NEWS.md
@@ -1,3 +1,9 @@
+# iplespBib 0.4.0
+
+* Team list is loaded from server now
+* Add `curl_verbose`, `curl_insecure` config options
+* Use new pagination feature to load records from the server
+
 # iplespBib 0.3.2
 
 * fix `recode_team()` to handle non existent team id, generating team label instead of returning NA